requirements.txt
fastapi
uvicorn[standard]
nltk
odfpy
tqdm
python-multipart
python-docx
PyPDF2
aiofiles
python-dotenv
requests